# docs/contracts/phase4a-openapi.json
{
  "components": {
    "schemas": {
...
      "DashboardProjection": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "concerns": {
            "$ref": "#/components/schemas/DashboardConcerns"
          },
          "country_name": {
            "title": "Country Name",
            "type": "string"
          },
          "goal": {
            "$ref": "#/components/schemas/GoalCard"
          },
          "government_form": {
            "title": "Government Form",
            "type": "string"
          },
          "map": {
            "$ref": "#/components/schemas/MapProjection"
          },
          "next_election_label": {
            "title": "Next Election Label",
            "type": "string"
          },
          "political_capital": {
            "$ref": "#/components/schemas/CapitalSummary"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "terminal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TerminalSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "turn": {
            "title": "Turn",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "turn",
          "country_name",
          "government_form",
          "next_election_label",
          "concerns",
          "political_capital",
          "goal",
          "map"
        ],
        "title": "DashboardProjection",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryDetailResponse": {
        "additionalProperties": false,
        "description": "Shape B -- historical detail. Same `TurnResultProjection` type as Shape A.",
        "properties": {
          "dashboardAsOfTurn": {
            "$ref": "#/components/schemas/DashboardProjection"
          },
          "turnResult": {
            "$ref": "#/components/schemas/TurnResultProjection"
          }
        },
        "required": [
          "turnResult",
          "dashboardAsOfTurn"
        ],
        "title": "HistoryDetailResponse",
        "type": "object"
      },
...
      "LoadRequest": {
        "additionalProperties": false,
        "properties": {
          "save_id": {
            "title": "Save Id",
            "type": "string"
          }
        },
        "required": [
          "save_id"
        ],
        "title": "LoadRequest",
        "type": "object"
      },
...
      "NewGameRequest": {
        "additionalProperties": false,
        "properties": {
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "NewGameRequest",
        "type": "object"
      },
...
      "SaveAsRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "SaveAsRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Local, loopback-only interface to the MANDATE simulation engine.",
    "title": "MANDATE local game API",
    "version": "0.15.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/game/history/{turn}": {
      "get": {
        "description": "One historical turn, rendered from its OWN stored report and state.\n\nNothing is recomputed from the current state: `entry_at(turn)` carries both\nthe report written when that turn resolved and the state it produced, and\n`dashboardAsOfTurn` is built from the latter.",
        "operationId": "get_history_detail_api_game_history__turn__get",
        "parameters": [
          {
            "in": "path",
            "name": "turn",
            "required": true,
            "schema": {
              "title": "Turn",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History Detail"
      }
    },
    "/api/game/load": {
      "post": {
        "description": "Load a save by ID.\n\nThe registry entry is never treated as proof of integrity: the save is read,\nparsed, version-checked and `validate_history`-checked, and only a save that\npasses all of that replaces the active session.",
        "operationId": "load_game_api_game_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Game"
      }
    },
...
    "/api/game/new": {
      "post": {
        "description": "Start a campaign.\n\nOrder (ADR 0014): build the save, allocate a UUID, write it atomically,\nupdate the index, and only then swap the active session -- so the in-memory\npointer never moves ahead of the disk.",
        "operationId": "create_game_api_game_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Game"
      }
    },
    "/api/game/preview": {
      "post": {
        "description": "Score a drafted proposal without changing anything.\n\nDeliberately does NOT take the mutation boundary: the frozen plan states\npreview uses \"never the session lock, never a write\", and taking it would let\na player drafting in one tab reject an unrelated resolve in another.\n\nInstead it captures `session.current_save` ONCE, at the top, and works\nentirely from that immutable reference -- so a concurrent resolve can never\nshow preview a mixture of pre- and post-mutation state. It observes one\ncomplete revision or the other.",
        "operationId": "preview_proposal_api_game_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Proposal"
      }
    },
    "/api/game/resolve": {
      "post": {
        "description": "Resolve one turn.\n\nThe order below is the authoritative one from ADR 0014, and the last two\nsteps are the ones that are easy to get subtly wrong:\n\n  1. Acquire mutation admission immediately, or reject `resolution_in_progress`.\n  2. Parse the client-echoed revision.\n  3. Compare it with the live state; reject a stale revision.\n  4. Build `DecisionSet` from the CLIENT's claimed turn/version, unsorted.\n  5. Call the engine.\n  6. Capture the returned immutable save in `new_save`.\n  7. Persist `new_save` atomically.\n  8. Swap `session.current_save` only after persistence succeeds.\n  9. Release the boundary -- on every path, success or failure.\n 10. Build the response from the CAPTURED `new_save`, never by re-reading\n     `session.current_save` after the boundary is released, which another\n     request could by then have advanced.\n\nA persistence failure leaves both the in-memory save and the on-disk save\nexactly as they were.",
        "operationId": "resolve_api_game_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve"
      }
    },
    "/api/game/save-as": {
      "post": {
        "description": "Checkpoint the current state under a NEW save id.\n\nA branch, not a rename: the running session keeps autosaving under its\noriginal id, which is left untouched.",
        "operationId": "save_as_api_game_save_as_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save As"
      }
    },
...
  }
}